change: protect footprints from delete

+++ fpmgr.php
@@ -25,26 +25,66 @@
 */
     include('lib.php');
     partdb_init();
+
+    /*
+     * In some cases a confirmation question has to be displayed.
+     */
+    $special_dialog = false;
     
+    /*
+     * this is the dispatcher ...
+     */
     $action = 'default';
     if ( isset( $_REQUEST["add"]))    { $action = 'add';}
     if ( isset( $_REQUEST["delete"])) { $action = 'delete';}
     if ( isset( $_REQUEST["rename"])) { $action = 'rename';}
-    
+   
+
     if ( $action == 'add')
     {
         $query = "INSERT INTO footprints (name) VALUES (". smart_escape($_REQUEST["new_footprint"]) .");";
         debug_print($query);
         mysql_query($query);
     }
-    
+   
+
     if ( $action == 'delete')
     {
-        // limit protects from runaway queries
-        $query = "DELETE FROM footprints WHERE id=". smart_escape($_REQUEST["footprint_sel"]) ." LIMIT 1;";
-        debug_print($query);
-        mysql_query($query);
+        $query = "SELECT (name) FROM parts WHERE id_footprint=". smart_escape($_REQUEST["footprint_sel"]) .";";
+        $r = mysql_query($query);
+        $ncol = mysql_num_rows ($r);
+        if ($ncol > 0)
+        {
+            $special_dialog = true;
+
+            // catch some example, where footprint is in use
+            for ($i = 0; ($i < $ncol) and ($i < 3); $i++)
+            {
+                $d         = mysql_fetch_row( $r);
+                $example[] = $d[0];
+            }
+            $example = implode( ', ', $example);
+
+            // give warning message
+            print "<html><body>".
+                "<div style=\"text-align:center;\">".
+                "<div style=\"color:red;font-size:x-large;\">Footprint kann nicht gel&ouml;scht werden!</div>";
+            if ($ncol == 1)
+                print "Es gibt noch 1 Teil, das diesen Footprint verwendet: ". $example;
+            else
+                print "Es gibt noch ". $ncol ." Teile, die diesen Footprint verwenden (z.B.: ". $example. ").";
+            print "<form method=\"get\" action=\"\">".
+                "<input type=\"submit\" value=\"OK\">".
+                "</form></div>".
+                "</body></html>";
+        }
+        else
+        {
+            $query = "DELETE FROM footprints WHERE id=". smart_escape($_REQUEST["footprint_sel"]) ." LIMIT 1;";
+            mysql_query($query);
+        }
     }
+
 
     if ( $action == 'rename')
     {
@@ -53,6 +93,11 @@
         mysql_query ($query);
     }
 
+    /*
+     * Don't show the default text when there's a msg.
+     */
+    if ($special_dialog == false)
+    {
 ?>
 <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
           "http://www.w3.org/TR/html4/loose.dtd">
@@ -127,3 +172,4 @@
 
 </body>
 </html>
+<?PHP } ?>
